ticket:24 Added login page
