+++ p2psafety_django/p2psafety/static/css/login.css
@@ -1,0 +1,9 @@
+body {
+    background-color: #eee;
+}
+
+.signin {
+    max-width: 400px;
+    padding: 15px;
+    margin: 0 auto;
+}
