+++ p2psafety_django/p2psafety/urls.py
@@ -1,20 +1,22 @@
+from django.conf import settings
 from django.conf.urls import patterns, include, url
+from django.conf.urls.static import static
 from django.contrib import admin
 from django.contrib.staticfiles.urls import staticfiles_urlpatterns
-from django.conf import settings
-from django.conf.urls.static import static
 
+from events.api import api_v1
 
+    
 admin.autodiscover()
 
 
 urlpatterns = patterns('',
-    # Examples:
-    # url(r'^$', 'p2psafety_django.views.home', name='home'),
-    # url(r'^blog/', include('blog.urls')),
-
-    url(r'^', include('events.urls')),
+    url(r'^', include('events.urls', namespace='events')),
     url(r'^admin/', include(admin.site.urls)),
+    url(r'^api/', include(api_v1.urls)),
+    url(r'^login/$', 'p2psafety.views.login'),
+    url(r'^logout/$', 'django.contrib.auth.views.logout', {'next_page': settings.LOGOUT_REDIRECT_URL}),
+    url(r'^accounts/', include('social.apps.django_app.urls', namespace='social')),
 )
 
 
@@ -24,4 +26,3 @@ if settings.DEBUG:
         settings.MEDIA_URL,
         document_root=settings.MEDIA_ROOT
     )
-
